@@ -4,27 +4,27 @@
   "private": true,
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
   "scripts": {
-    "postinstall": "rm -rf node_modules/react-navigation/{node_modules,examples}",
     "start": "react-native-scripts start",
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
     "ios": "react-native-scripts ios",
-    "test": "node node_modules/jest/bin/jest.js"
+    "test": "node node_modules/jest/bin/jest.js && flow"
   },
   "dependencies": {
-    "expo": "^23.0.0",
-    "react": "16.0.0",
-    "react-native": "^0.50.3",
-    "react-navigation": "file:../.."
+    "expo": "^26.0.0",
+    "react": "16.3.0-alpha.1",
+    "react-native": "^0.54.0",
+    "react-native-iphone-x-helper": "^1.0.2",
+    "react-navigation": "link:../.."
   },
   "devDependencies": {
     "babel-jest": "^21.0.0",
-    "flow-bin": "^0.56.0",
+    "babel-plugin-transform-remove-console": "^6.9.0",
+    "flow-bin": "^0.61.0",
     "jest": "^21.0.1",
-    "jest-expo": "^23.0.0",
-    "react-addons-test-utils": "16.0.0-alpha.3",
+    "jest-expo": "^26.0.0",
     "react-native-scripts": "^1.5.0",
-    "react-test-renderer": "16.0.0-alpha.12"
+    "react-test-renderer": "16.3.0-alpha.1"
   },
   "jest": {
     "preset": "jest-expo",
