@@ -4,7 +4,6 @@
   "private": true,
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
   "scripts": {
-    "postinstall": "rm -rf node_modules/react-navigation/{node_modules,examples}",
     "start": "react-native-scripts start",
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
@@ -22,19 +21,20 @@
     ]
   },
   "dependencies": {
-    "expo": "^23.0.0",
+    "expo": "^25.0.0",
     "prop-types": "^15.5.10",
-    "react": "16.0.0",
-    "react-native": "^0.50.3",
+    "react": "16.2.0",
+    "react-native": "^0.52.0",
+    "react-navigation": "link:../..",
+    "react-navigation-redux-helpers": "^1.0.3",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2"
   },
   "devDependencies": {
     "babel-jest": "^21.0.0",
     "jest": "^21.0.1",
-    "jest-expo": "^23.0.0",
+    "jest-expo": "^25.1.0",
     "react-native-scripts": "^1.3.1",
-    "react-navigation": "file:../..",
-    "react-test-renderer": "16.0.0-alpha.12"
+    "react-test-renderer": "16.0.0"
   }
 }
