@@ -1,5 +1,5 @@
-// flow-typed signature: 5964a245d3129ecd50da1f74a2f86454
-// flow-typed version: <<STUB>>/babel-jest_v^21.0.0/flow_v0.53.0
+// flow-typed signature: ee7b4028012cfdcd83cc6541a39b2282
+// flow-typed version: <<STUB>>/babel-jest_v^21.0.0/flow_v0.61.0
 
 /**
  * This is an autogenerated libdef stub for:
