@@ -1,5 +1,5 @@
-// flow-typed signature: 4fb1f6897647ab1ad25d819aa06d2d6c
-// flow-typed version: <<STUB>>/expo_v^22.0.0/flow_v0.53.0
+// flow-typed signature: c1ecfe71fbb86c3602b7da3d7c141df1
+// flow-typed version: <<STUB>>/expo_v^24.0.2/flow_v0.61.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -23,6 +23,18 @@ declare module 'expo' {
  * needed.
  */
 declare module 'expo/AppEntry' {
+  declare module.exports: any;
+}
+
+declare module 'expo/flow/metro-bundler' {
+  declare module.exports: any;
+}
+
+declare module 'expo/flow/react-native-gesture-handler' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/__mocks__/Constants-development' {
   declare module.exports: any;
 }
 
@@ -50,6 +62,10 @@ declare module 'expo/src/__tests__/Facebook-test' {
   declare module.exports: any;
 }
 
+declare module 'expo/src/__tests__/Font-test' {
+  declare module.exports: any;
+}
+
 declare module 'expo/src/__tests__/Location-test' {
   declare module.exports: any;
 }
@@ -70,6 +86,22 @@ declare module 'expo/src/__tests__/WebBrowser-test' {
   declare module.exports: any;
 }
 
+declare module 'expo/src/admob/RNAdMobBanner' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/admob/RNAdMobInterstitial' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/admob/RNAdMobRewarded' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/admob/RNPublisherBanner' {
+  declare module.exports: any;
+}
+
 declare module 'expo/src/Amplitude' {
   declare module.exports: any;
 }
@@ -78,23 +110,7 @@ declare module 'expo/src/apisAreAvailable' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/AppLoading' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/AppLoadingNativeWrapper.android' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/AppLoadingNativeWrapper.ios' {
-  declare module.exports: any;
-}
-
 declare module 'expo/src/Asset' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/Audio' {
   declare module.exports: any;
 }
 
@@ -102,19 +118,31 @@ declare module 'expo/src/AuthSession' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/AV' {
+declare module 'expo/src/av/Audio' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/BarCodeScanner' {
+declare module 'expo/src/av/Audio/Recording' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/BlurView.android' {
+declare module 'expo/src/av/Audio/Sound' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/BlurView.ios' {
+declare module 'expo/src/av/AV' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/av/Video' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/BarCodeScanner.android' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/BarCodeScanner.ios' {
   declare module.exports: any;
 }
 
@@ -123,6 +151,10 @@ declare module 'expo/src/Brightness' {
 }
 
 declare module 'expo/src/Camera' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/CameraBasedBarCodeScanner' {
   declare module.exports: any;
 }
 
@@ -142,6 +174,34 @@ declare module 'expo/src/DocumentPicker' {
   declare module.exports: any;
 }
 
+declare module 'expo/src/effects/BlurView.android' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/effects/BlurView.ios' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/effects/LinearGradient.android' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/effects/LinearGradient.ios' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/environment/__tests__/validate-test' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/environment/logging' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/environment/validate' {
+  declare module.exports: any;
+}
+
 declare module 'expo/src/ErrorRecovery' {
   declare module.exports: any;
 }
@@ -150,31 +210,35 @@ declare module 'expo/src/Expo' {
   declare module.exports: any;
 }
 
+declare module 'expo/src/facebook-ads/AdSettings' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/facebook-ads/BannerViewManager' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/facebook-ads/index' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/facebook-ads/InterstitialAdManager' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/facebook-ads/NativeAdsManager' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/facebook-ads/withNativeAd' {
+  declare module.exports: any;
+}
+
 declare module 'expo/src/Facebook' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/FacebookAds/AdSettings' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/FacebookAds/BannerViewManager' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/FacebookAds/index' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/FacebookAds/InterstitialAdManager' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/FacebookAds/NativeAdsManager' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/FacebookAds/withNativeAd' {
+declare module 'expo/src/FaceDetector' {
   declare module.exports: any;
 }
 
@@ -202,6 +266,10 @@ declare module 'expo/src/Icon' {
   declare module.exports: any;
 }
 
+declare module 'expo/src/ImageManipulator' {
+  declare module.exports: any;
+}
+
 declare module 'expo/src/ImagePicker' {
   declare module.exports: any;
 }
@@ -214,11 +282,23 @@ declare module 'expo/src/KeepAwake' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/LegacyAsyncStorage.android' {
+declare module 'expo/src/launch/AppLoading' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/LegacyAsyncStorage.ios' {
+declare module 'expo/src/launch/AppLoadingNativeWrapper.android' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/launch/AppLoadingNativeWrapper.ios' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/launch/registerRootComponent' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/launch/RootErrorBoundary' {
   declare module.exports: any;
 }
 
@@ -226,39 +306,67 @@ declare module 'expo/src/lib/Queue' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/LinearGradient.android' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/LinearGradient.ios' {
-  declare module.exports: any;
-}
-
 declare module 'expo/src/Location' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/Logs' {
+declare module 'expo/src/logs/__tests__/Logs-test' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/Modal/Modal' {
+declare module 'expo/src/logs/__tests__/LogSerialization-test' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/Modal/ModalHost' {
+declare module 'expo/src/logs/__tests__/RemoteConsole-test' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/Modal/ModalImplementation' {
+declare module 'expo/src/logs/__tests__/RemoteLogging-test' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/Modal/PureContainer' {
+declare module 'expo/src/logs/__tests__/RemoteLogs-test' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/logs/Logs' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/logs/LogSerialization' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/logs/RemoteConsole' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/logs/RemoteLogging' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/modal/Modal' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/modal/ModalHost' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/modal/ModalImplementation' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/modal/PureContainer' {
   declare module.exports: any;
 }
 
 declare module 'expo/src/Notifications' {
+  declare module.exports: any;
+}
+
+declare module 'expo/src/OldBarCodeScanner' {
   declare module.exports: any;
 }
 
@@ -271,30 +379,6 @@ declare module 'expo/src/Pedometer' {
 }
 
 declare module 'expo/src/Permissions' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/registerRootComponent' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/RNAdMobBanner' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/RNAdMobInterstitial' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/RNAdMobRewarded' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/RNPublisherBanner' {
-  declare module.exports: any;
-}
-
-declare module 'expo/src/RootErrorBoundary' {
   declare module.exports: any;
 }
 
@@ -374,11 +458,11 @@ declare module 'expo/src/takeSnapshotAsync' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/Util' {
+declare module 'expo/src/timer/polyfillNextTick' {
   declare module.exports: any;
 }
 
-declare module 'expo/src/Video' {
+declare module 'expo/src/Util' {
   declare module.exports: any;
 }
 
@@ -398,6 +482,15 @@ declare module 'expo/tools/LogReporter' {
 declare module 'expo/AppEntry.js' {
   declare module.exports: $Exports<'expo/AppEntry'>;
 }
+declare module 'expo/flow/metro-bundler.js' {
+  declare module.exports: $Exports<'expo/flow/metro-bundler'>;
+}
+declare module 'expo/flow/react-native-gesture-handler.js' {
+  declare module.exports: $Exports<'expo/flow/react-native-gesture-handler'>;
+}
+declare module 'expo/src/__mocks__/Constants-development.js' {
+  declare module.exports: $Exports<'expo/src/__mocks__/Constants-development'>;
+}
 declare module 'expo/src/__tests__/Amplitude-test.js' {
   declare module.exports: $Exports<'expo/src/__tests__/Amplitude-test'>;
 }
@@ -416,6 +509,9 @@ declare module 'expo/src/__tests__/Expo-test.js' {
 declare module 'expo/src/__tests__/Facebook-test.js' {
   declare module.exports: $Exports<'expo/src/__tests__/Facebook-test'>;
 }
+declare module 'expo/src/__tests__/Font-test.js' {
+  declare module.exports: $Exports<'expo/src/__tests__/Font-test'>;
+}
 declare module 'expo/src/__tests__/Location-test.js' {
   declare module.exports: $Exports<'expo/src/__tests__/Location-test'>;
 }
@@ -431,47 +527,59 @@ declare module 'expo/src/__tests__/Segment-test.js' {
 declare module 'expo/src/__tests__/WebBrowser-test.js' {
   declare module.exports: $Exports<'expo/src/__tests__/WebBrowser-test'>;
 }
+declare module 'expo/src/admob/RNAdMobBanner.js' {
+  declare module.exports: $Exports<'expo/src/admob/RNAdMobBanner'>;
+}
+declare module 'expo/src/admob/RNAdMobInterstitial.js' {
+  declare module.exports: $Exports<'expo/src/admob/RNAdMobInterstitial'>;
+}
+declare module 'expo/src/admob/RNAdMobRewarded.js' {
+  declare module.exports: $Exports<'expo/src/admob/RNAdMobRewarded'>;
+}
+declare module 'expo/src/admob/RNPublisherBanner.js' {
+  declare module.exports: $Exports<'expo/src/admob/RNPublisherBanner'>;
+}
 declare module 'expo/src/Amplitude.js' {
   declare module.exports: $Exports<'expo/src/Amplitude'>;
 }
 declare module 'expo/src/apisAreAvailable.js' {
   declare module.exports: $Exports<'expo/src/apisAreAvailable'>;
 }
-declare module 'expo/src/AppLoading.js' {
-  declare module.exports: $Exports<'expo/src/AppLoading'>;
-}
-declare module 'expo/src/AppLoadingNativeWrapper.android.js' {
-  declare module.exports: $Exports<'expo/src/AppLoadingNativeWrapper.android'>;
-}
-declare module 'expo/src/AppLoadingNativeWrapper.ios.js' {
-  declare module.exports: $Exports<'expo/src/AppLoadingNativeWrapper.ios'>;
-}
 declare module 'expo/src/Asset.js' {
   declare module.exports: $Exports<'expo/src/Asset'>;
-}
-declare module 'expo/src/Audio.js' {
-  declare module.exports: $Exports<'expo/src/Audio'>;
 }
 declare module 'expo/src/AuthSession.js' {
   declare module.exports: $Exports<'expo/src/AuthSession'>;
 }
-declare module 'expo/src/AV.js' {
-  declare module.exports: $Exports<'expo/src/AV'>;
+declare module 'expo/src/av/Audio.js' {
+  declare module.exports: $Exports<'expo/src/av/Audio'>;
 }
-declare module 'expo/src/BarCodeScanner.js' {
-  declare module.exports: $Exports<'expo/src/BarCodeScanner'>;
+declare module 'expo/src/av/Audio/Recording.js' {
+  declare module.exports: $Exports<'expo/src/av/Audio/Recording'>;
 }
-declare module 'expo/src/BlurView.android.js' {
-  declare module.exports: $Exports<'expo/src/BlurView.android'>;
+declare module 'expo/src/av/Audio/Sound.js' {
+  declare module.exports: $Exports<'expo/src/av/Audio/Sound'>;
 }
-declare module 'expo/src/BlurView.ios.js' {
-  declare module.exports: $Exports<'expo/src/BlurView.ios'>;
+declare module 'expo/src/av/AV.js' {
+  declare module.exports: $Exports<'expo/src/av/AV'>;
+}
+declare module 'expo/src/av/Video.js' {
+  declare module.exports: $Exports<'expo/src/av/Video'>;
+}
+declare module 'expo/src/BarCodeScanner.android.js' {
+  declare module.exports: $Exports<'expo/src/BarCodeScanner.android'>;
+}
+declare module 'expo/src/BarCodeScanner.ios.js' {
+  declare module.exports: $Exports<'expo/src/BarCodeScanner.ios'>;
 }
 declare module 'expo/src/Brightness.js' {
   declare module.exports: $Exports<'expo/src/Brightness'>;
 }
 declare module 'expo/src/Camera.js' {
   declare module.exports: $Exports<'expo/src/Camera'>;
+}
+declare module 'expo/src/CameraBasedBarCodeScanner.js' {
+  declare module.exports: $Exports<'expo/src/CameraBasedBarCodeScanner'>;
 }
 declare module 'expo/src/Constants.js' {
   declare module.exports: $Exports<'expo/src/Constants'>;
@@ -485,32 +593,56 @@ declare module 'expo/src/DangerZone.js' {
 declare module 'expo/src/DocumentPicker.js' {
   declare module.exports: $Exports<'expo/src/DocumentPicker'>;
 }
+declare module 'expo/src/effects/BlurView.android.js' {
+  declare module.exports: $Exports<'expo/src/effects/BlurView.android'>;
+}
+declare module 'expo/src/effects/BlurView.ios.js' {
+  declare module.exports: $Exports<'expo/src/effects/BlurView.ios'>;
+}
+declare module 'expo/src/effects/LinearGradient.android.js' {
+  declare module.exports: $Exports<'expo/src/effects/LinearGradient.android'>;
+}
+declare module 'expo/src/effects/LinearGradient.ios.js' {
+  declare module.exports: $Exports<'expo/src/effects/LinearGradient.ios'>;
+}
+declare module 'expo/src/environment/__tests__/validate-test.js' {
+  declare module.exports: $Exports<'expo/src/environment/__tests__/validate-test'>;
+}
+declare module 'expo/src/environment/logging.js' {
+  declare module.exports: $Exports<'expo/src/environment/logging'>;
+}
+declare module 'expo/src/environment/validate.js' {
+  declare module.exports: $Exports<'expo/src/environment/validate'>;
+}
 declare module 'expo/src/ErrorRecovery.js' {
   declare module.exports: $Exports<'expo/src/ErrorRecovery'>;
 }
 declare module 'expo/src/Expo.js' {
   declare module.exports: $Exports<'expo/src/Expo'>;
 }
+declare module 'expo/src/facebook-ads/AdSettings.js' {
+  declare module.exports: $Exports<'expo/src/facebook-ads/AdSettings'>;
+}
+declare module 'expo/src/facebook-ads/BannerViewManager.js' {
+  declare module.exports: $Exports<'expo/src/facebook-ads/BannerViewManager'>;
+}
+declare module 'expo/src/facebook-ads/index.js' {
+  declare module.exports: $Exports<'expo/src/facebook-ads/index'>;
+}
+declare module 'expo/src/facebook-ads/InterstitialAdManager.js' {
+  declare module.exports: $Exports<'expo/src/facebook-ads/InterstitialAdManager'>;
+}
+declare module 'expo/src/facebook-ads/NativeAdsManager.js' {
+  declare module.exports: $Exports<'expo/src/facebook-ads/NativeAdsManager'>;
+}
+declare module 'expo/src/facebook-ads/withNativeAd.js' {
+  declare module.exports: $Exports<'expo/src/facebook-ads/withNativeAd'>;
+}
 declare module 'expo/src/Facebook.js' {
   declare module.exports: $Exports<'expo/src/Facebook'>;
 }
-declare module 'expo/src/FacebookAds/AdSettings.js' {
-  declare module.exports: $Exports<'expo/src/FacebookAds/AdSettings'>;
-}
-declare module 'expo/src/FacebookAds/BannerViewManager.js' {
-  declare module.exports: $Exports<'expo/src/FacebookAds/BannerViewManager'>;
-}
-declare module 'expo/src/FacebookAds/index.js' {
-  declare module.exports: $Exports<'expo/src/FacebookAds/index'>;
-}
-declare module 'expo/src/FacebookAds/InterstitialAdManager.js' {
-  declare module.exports: $Exports<'expo/src/FacebookAds/InterstitialAdManager'>;
-}
-declare module 'expo/src/FacebookAds/NativeAdsManager.js' {
-  declare module.exports: $Exports<'expo/src/FacebookAds/NativeAdsManager'>;
-}
-declare module 'expo/src/FacebookAds/withNativeAd.js' {
-  declare module.exports: $Exports<'expo/src/FacebookAds/withNativeAd'>;
+declare module 'expo/src/FaceDetector.js' {
+  declare module.exports: $Exports<'expo/src/FaceDetector'>;
 }
 declare module 'expo/src/FileSystem.js' {
   declare module.exports: $Exports<'expo/src/FileSystem'>;
@@ -530,6 +662,9 @@ declare module 'expo/src/Google.js' {
 declare module 'expo/src/Icon.js' {
   declare module.exports: $Exports<'expo/src/Icon'>;
 }
+declare module 'expo/src/ImageManipulator.js' {
+  declare module.exports: $Exports<'expo/src/ImageManipulator'>;
+}
 declare module 'expo/src/ImagePicker.js' {
   declare module.exports: $Exports<'expo/src/ImagePicker'>;
 }
@@ -539,41 +674,71 @@ declare module 'expo/src/IntentLauncherAndroid.js' {
 declare module 'expo/src/KeepAwake.js' {
   declare module.exports: $Exports<'expo/src/KeepAwake'>;
 }
-declare module 'expo/src/LegacyAsyncStorage.android.js' {
-  declare module.exports: $Exports<'expo/src/LegacyAsyncStorage.android'>;
+declare module 'expo/src/launch/AppLoading.js' {
+  declare module.exports: $Exports<'expo/src/launch/AppLoading'>;
 }
-declare module 'expo/src/LegacyAsyncStorage.ios.js' {
-  declare module.exports: $Exports<'expo/src/LegacyAsyncStorage.ios'>;
+declare module 'expo/src/launch/AppLoadingNativeWrapper.android.js' {
+  declare module.exports: $Exports<'expo/src/launch/AppLoadingNativeWrapper.android'>;
+}
+declare module 'expo/src/launch/AppLoadingNativeWrapper.ios.js' {
+  declare module.exports: $Exports<'expo/src/launch/AppLoadingNativeWrapper.ios'>;
+}
+declare module 'expo/src/launch/registerRootComponent.js' {
+  declare module.exports: $Exports<'expo/src/launch/registerRootComponent'>;
+}
+declare module 'expo/src/launch/RootErrorBoundary.js' {
+  declare module.exports: $Exports<'expo/src/launch/RootErrorBoundary'>;
 }
 declare module 'expo/src/lib/Queue.js' {
   declare module.exports: $Exports<'expo/src/lib/Queue'>;
 }
-declare module 'expo/src/LinearGradient.android.js' {
-  declare module.exports: $Exports<'expo/src/LinearGradient.android'>;
-}
-declare module 'expo/src/LinearGradient.ios.js' {
-  declare module.exports: $Exports<'expo/src/LinearGradient.ios'>;
-}
 declare module 'expo/src/Location.js' {
   declare module.exports: $Exports<'expo/src/Location'>;
 }
-declare module 'expo/src/Logs.js' {
-  declare module.exports: $Exports<'expo/src/Logs'>;
+declare module 'expo/src/logs/__tests__/Logs-test.js' {
+  declare module.exports: $Exports<'expo/src/logs/__tests__/Logs-test'>;
 }
-declare module 'expo/src/Modal/Modal.js' {
-  declare module.exports: $Exports<'expo/src/Modal/Modal'>;
+declare module 'expo/src/logs/__tests__/LogSerialization-test.js' {
+  declare module.exports: $Exports<'expo/src/logs/__tests__/LogSerialization-test'>;
 }
-declare module 'expo/src/Modal/ModalHost.js' {
-  declare module.exports: $Exports<'expo/src/Modal/ModalHost'>;
+declare module 'expo/src/logs/__tests__/RemoteConsole-test.js' {
+  declare module.exports: $Exports<'expo/src/logs/__tests__/RemoteConsole-test'>;
 }
-declare module 'expo/src/Modal/ModalImplementation.js' {
-  declare module.exports: $Exports<'expo/src/Modal/ModalImplementation'>;
+declare module 'expo/src/logs/__tests__/RemoteLogging-test.js' {
+  declare module.exports: $Exports<'expo/src/logs/__tests__/RemoteLogging-test'>;
 }
-declare module 'expo/src/Modal/PureContainer.js' {
-  declare module.exports: $Exports<'expo/src/Modal/PureContainer'>;
+declare module 'expo/src/logs/__tests__/RemoteLogs-test.js' {
+  declare module.exports: $Exports<'expo/src/logs/__tests__/RemoteLogs-test'>;
+}
+declare module 'expo/src/logs/Logs.js' {
+  declare module.exports: $Exports<'expo/src/logs/Logs'>;
+}
+declare module 'expo/src/logs/LogSerialization.js' {
+  declare module.exports: $Exports<'expo/src/logs/LogSerialization'>;
+}
+declare module 'expo/src/logs/RemoteConsole.js' {
+  declare module.exports: $Exports<'expo/src/logs/RemoteConsole'>;
+}
+declare module 'expo/src/logs/RemoteLogging.js' {
+  declare module.exports: $Exports<'expo/src/logs/RemoteLogging'>;
+}
+declare module 'expo/src/modal/Modal.js' {
+  declare module.exports: $Exports<'expo/src/modal/Modal'>;
+}
+declare module 'expo/src/modal/ModalHost.js' {
+  declare module.exports: $Exports<'expo/src/modal/ModalHost'>;
+}
+declare module 'expo/src/modal/ModalImplementation.js' {
+  declare module.exports: $Exports<'expo/src/modal/ModalImplementation'>;
+}
+declare module 'expo/src/modal/PureContainer.js' {
+  declare module.exports: $Exports<'expo/src/modal/PureContainer'>;
 }
 declare module 'expo/src/Notifications.js' {
   declare module.exports: $Exports<'expo/src/Notifications'>;
+}
+declare module 'expo/src/OldBarCodeScanner.js' {
+  declare module.exports: $Exports<'expo/src/OldBarCodeScanner'>;
 }
 declare module 'expo/src/Payments.js' {
   declare module.exports: $Exports<'expo/src/Payments'>;
@@ -583,24 +748,6 @@ declare module 'expo/src/Pedometer.js' {
 }
 declare module 'expo/src/Permissions.js' {
   declare module.exports: $Exports<'expo/src/Permissions'>;
-}
-declare module 'expo/src/registerRootComponent.js' {
-  declare module.exports: $Exports<'expo/src/registerRootComponent'>;
-}
-declare module 'expo/src/RNAdMobBanner.js' {
-  declare module.exports: $Exports<'expo/src/RNAdMobBanner'>;
-}
-declare module 'expo/src/RNAdMobInterstitial.js' {
-  declare module.exports: $Exports<'expo/src/RNAdMobInterstitial'>;
-}
-declare module 'expo/src/RNAdMobRewarded.js' {
-  declare module.exports: $Exports<'expo/src/RNAdMobRewarded'>;
-}
-declare module 'expo/src/RNPublisherBanner.js' {
-  declare module.exports: $Exports<'expo/src/RNPublisherBanner'>;
-}
-declare module 'expo/src/RootErrorBoundary.js' {
-  declare module.exports: $Exports<'expo/src/RootErrorBoundary'>;
 }
 declare module 'expo/src/ScreenOrientation.js' {
   declare module.exports: $Exports<'expo/src/ScreenOrientation'>;
@@ -659,11 +806,11 @@ declare module 'expo/src/Svg.js' {
 declare module 'expo/src/takeSnapshotAsync.js' {
   declare module.exports: $Exports<'expo/src/takeSnapshotAsync'>;
 }
+declare module 'expo/src/timer/polyfillNextTick.js' {
+  declare module.exports: $Exports<'expo/src/timer/polyfillNextTick'>;
+}
 declare module 'expo/src/Util.js' {
   declare module.exports: $Exports<'expo/src/Util'>;
-}
-declare module 'expo/src/Video.js' {
-  declare module.exports: $Exports<'expo/src/Video'>;
 }
 declare module 'expo/src/WebBrowser.js' {
   declare module.exports: $Exports<'expo/src/WebBrowser'>;
