@@ -1,5 +1,5 @@
-// flow-typed signature: 72ab5b9c7dcb88b82de627702a0db27a
-// flow-typed version: <<STUB>>/jest-expo_v^22.0.0/flow_v0.53.0
+// flow-typed signature: b1e3826a494aecf4ff482336321b54bd
+// flow-typed version: <<STUB>>/jest-expo_v^24.0.0/flow_v0.61.0
 
 /**
  * This is an autogenerated libdef stub for:
