@@ -1,5 +1,5 @@
-// flow-typed signature: bd52927dcfdd08c54751906ebaf7cfa1
-// flow-typed version: <<STUB>>/react-native-scripts_v^1.3.1/flow_v0.53.0
+// flow-typed signature: 42d8cd2e84e894c09d28f891fef9f01d
+// flow-typed version: <<STUB>>/react-native-scripts_v^1.5.0/flow_v0.61.0
 
 /**
  * This is an autogenerated libdef stub for:
