@@ -1,8 +1,4 @@
-/* @flow */
-
 import invariant from './utils/invariant';
-
-import type { NavigationRoute, NavigationState } from './TypeDefinition';
 
 /**
  * Utilities to perform atomic operation with navigate state and routes.
@@ -16,24 +12,24 @@ const StateUtils = {
   /**
    * Gets a route by key. If the route isn't found, returns `null`.
    */
-  get(state: NavigationState, key: string): ?NavigationRoute {
-    return state.routes.find((route: *) => route.key === key) || null;
+  get(state, key) {
+    return state.routes.find(route => route.key === key) || null;
   },
 
   /**
    * Returns the first index at which a given route's key can be found in the
    * routes of the navigation state, or -1 if it is not present.
    */
-  indexOf(state: NavigationState, key: string): number {
-    return state.routes.map((route: *) => route.key).indexOf(key);
+  indexOf(state, key) {
+    return state.routes.map(route => route.key).indexOf(key);
   },
 
   /**
    * Returns `true` at which a given route's key can be found in the
    * routes of the navigation state.
    */
-  has(state: NavigationState, key: string): boolean {
-    return !!state.routes.some((route: *) => route.key === key);
+  has(state, key) {
+    return !!state.routes.some(route => route.key === key);
   },
 
   /**
@@ -41,7 +37,7 @@ const StateUtils = {
    * Note that this moves the index to the positon to where the last route in the
    * stack is at.
    */
-  push(state: NavigationState, route: NavigationRoute): NavigationState {
+  push(state, route) {
     invariant(
       StateUtils.indexOf(state, route.key) === -1,
       'should not push route with duplicated key %s',
@@ -63,7 +59,7 @@ const StateUtils = {
    * Note that this moves the index to the positon to where the last route in the
    * stack is at.
    */
-  pop(state: NavigationState): NavigationState {
+  pop(state) {
     if (state.index <= 0) {
       // [Note]: Over-popping does not throw error. Instead, it will be no-op.
       return state;
@@ -79,7 +75,7 @@ const StateUtils = {
   /**
    * Sets the focused route of the navigation state by index.
    */
-  jumpToIndex(state: NavigationState, index: number): NavigationState {
+  jumpToIndex(state, index) {
     if (index === state.index) {
       return state;
     }
@@ -95,7 +91,7 @@ const StateUtils = {
   /**
    * Sets the focused route of the navigation state by key.
    */
-  jumpTo(state: NavigationState, key: string): NavigationState {
+  jumpTo(state, key) {
     const index = StateUtils.indexOf(state, key);
     return StateUtils.jumpToIndex(state, index);
   },
@@ -103,7 +99,7 @@ const StateUtils = {
   /**
    * Sets the focused route to the previous route.
    */
-  back(state: NavigationState): NavigationState {
+  back(state) {
     const index = state.index - 1;
     const route = state.routes[index];
     return route ? StateUtils.jumpToIndex(state, index) : state;
@@ -112,7 +108,7 @@ const StateUtils = {
   /**
    * Sets the focused route to the next route.
    */
-  forward(state: NavigationState): NavigationState {
+  forward(state) {
     const index = state.index + 1;
     const route = state.routes[index];
     return route ? StateUtils.jumpToIndex(state, index) : state;
@@ -123,11 +119,7 @@ const StateUtils = {
    * Note that this moves the index to the positon to where the new route in the
    * stack is at.
    */
-  replaceAt(
-    state: NavigationState,
-    key: string,
-    route: NavigationRoute
-  ): NavigationState {
+  replaceAt(state, key, route) {
     const index = StateUtils.indexOf(state, key);
     return StateUtils.replaceAtIndex(state, index, route);
   },
@@ -137,11 +129,7 @@ const StateUtils = {
    * Note that this moves the index to the positon to where the new route in the
    * stack is at.
    */
-  replaceAtIndex(
-    state: NavigationState,
-    index: number,
-    route: NavigationRoute
-  ): NavigationState {
+  replaceAtIndex(state, index, route) {
     invariant(
       !!state.routes[index],
       'invalid index %s for replacing route %s',
@@ -168,20 +156,16 @@ const StateUtils = {
    * Note that this moves the index to the positon to where the last route in the
    * stack is at if the param `index` isn't provided.
    */
-  reset(
-    state: NavigationState,
-    routes: Array<NavigationRoute>,
-    index?: number
-  ): NavigationState {
+  reset(state, routes, index) {
     invariant(
       routes.length && Array.isArray(routes),
       'invalid routes to replace'
     );
 
-    const nextIndex: number = index === undefined ? routes.length - 1 : index;
+    const nextIndex = index === undefined ? routes.length - 1 : index;
 
     if (state.routes.length === routes.length && state.index === nextIndex) {
-      const compare = (route: *, ii: *) => routes[ii] === route;
+      const compare = (route, ii) => routes[ii] === route;
       if (state.routes.every(compare)) {
         return state;
       }
