@@ -1,7 +1,9 @@
-/* @flow */
-
 import NavigationActions from '../NavigationActions';
 import addNavigationHelpers from '../addNavigationHelpers';
+
+const dummyEventSubscriber = (name: string, handler: (*) => void) => ({
+  remove: () => {},
+});
 
 describe('addNavigationHelpers', () => {
   it('handles Back action', () => {
@@ -12,6 +14,7 @@ describe('addNavigationHelpers', () => {
       addNavigationHelpers({
         state: { key: 'A', routeName: 'Home' },
         dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
       }).goBack('A')
     ).toEqual(true);
     expect(mockedDispatch).toBeCalledWith({
@@ -29,6 +32,7 @@ describe('addNavigationHelpers', () => {
       addNavigationHelpers({
         state: { routeName: 'Home' },
         dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
       }).goBack()
     ).toEqual(true);
     expect(mockedDispatch).toBeCalledWith({ type: NavigationActions.BACK });
@@ -43,6 +47,7 @@ describe('addNavigationHelpers', () => {
       addNavigationHelpers({
         state: { routeName: 'Home' },
         dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
       }).navigate('Profile', { name: 'Matt' })
     ).toEqual(true);
     expect(mockedDispatch).toBeCalledWith({
@@ -61,6 +66,7 @@ describe('addNavigationHelpers', () => {
       addNavigationHelpers({
         state: { key: 'B', routeName: 'Settings' },
         dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
       }).setParams({ notificationsEnabled: 'yes' })
     ).toEqual(true);
     expect(mockedDispatch).toBeCalledWith({
@@ -69,5 +75,44 @@ describe('addNavigationHelpers', () => {
       params: { notificationsEnabled: 'yes' },
     });
     expect(mockedDispatch.mock.calls.length).toBe(1);
+  });
+
+  it('handles GetParams action', () => {
+    const mockedDispatch = jest
+      .fn(() => false)
+      .mockImplementationOnce(() => true);
+    expect(
+      addNavigationHelpers({
+        state: { key: 'B', routeName: 'Settings', params: { name: 'Peter' } },
+        dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
+      }).getParam('name', 'Brent')
+    ).toEqual('Peter');
+  });
+
+  it('handles GetParams action with default param', () => {
+    const mockedDispatch = jest
+      .fn(() => false)
+      .mockImplementationOnce(() => true);
+    expect(
+      addNavigationHelpers({
+        state: { key: 'B', routeName: 'Settings' },
+        dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
+      }).getParam('name', 'Brent')
+    ).toEqual('Brent');
+  });
+
+  it('handles GetParams action with param value as null', () => {
+    const mockedDispatch = jest
+      .fn(() => false)
+      .mockImplementationOnce(() => true);
+    expect(
+      addNavigationHelpers({
+        state: { key: 'B', routeName: 'Settings', params: { name: null } },
+        dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
+      }).getParam('name')
+    ).toEqual(null);
   });
 });
