@@ -1,26 +1,13 @@
-/**
- * @flow
- *
- * Helpers for navigation.
- */
-
-import type {
-  NavigationProp,
-  NavigationParams,
-  NavigationScreenProp,
-  NavigationNavigateAction,
-} from './TypeDefinition';
+/* Helpers for navigation */
 
 import NavigationActions from './NavigationActions';
 import invariant from './utils/invariant';
 
-export default function<S: {}>(
-  navigation: NavigationProp<S>
-): NavigationScreenProp<S> {
+export default function(navigation) {
   return {
     ...navigation,
-    goBack: (key?: ?string): boolean => {
-      let actualizedKey: ?string = key;
+    goBack: key => {
+      let actualizedKey = key;
       if (key === undefined && navigation.state.key) {
         invariant(
           typeof navigation.state.key === 'string',
@@ -32,20 +19,40 @@ export default function<S: {}>(
         NavigationActions.back({ key: actualizedKey })
       );
     },
-    navigate: (
-      routeName: string,
-      params?: NavigationParams,
-      action?: NavigationNavigateAction
-    ): boolean =>
+    navigate: (navigateTo, params, action) => {
+      if (typeof navigateTo === 'string') {
+        return navigation.dispatch(
+          NavigationActions.navigate({ routeName: navigateTo, params, action })
+        );
+      }
+      invariant(
+        typeof navigateTo === 'object',
+        'Must navigateTo an object or a string'
+      );
+      invariant(
+        params == null,
+        'Params must not be provided to .navigate() when specifying an object'
+      );
+      invariant(
+        action == null,
+        'Child action must not be provided to .navigate() when specifying an object'
+      );
+      return navigation.dispatch(NavigationActions.navigate(navigateTo));
+    },
+    pop: (n, params) =>
       navigation.dispatch(
-        NavigationActions.navigate({ routeName, params, action })
+        NavigationActions.pop({ n, immediate: params && params.immediate })
+      ),
+    popToTop: params =>
+      navigation.dispatch(
+        NavigationActions.popToTop({ immediate: params && params.immediate })
       ),
     /**
      * For updating current route params. For example the nav bar title and
      * buttons are based on the route params.
      * This means `setParams` can be used to update nav bar for example.
      */
-    setParams: (params: NavigationParams): boolean => {
+    setParams: params => {
       invariant(
         navigation.state.key && typeof navigation.state.key === 'string',
         'setParams cannot be called by root navigator'
@@ -53,5 +60,30 @@ export default function<S: {}>(
       const key = navigation.state.key;
       return navigation.dispatch(NavigationActions.setParams({ params, key }));
     },
+
+    getParam: (paramName, defaultValue) => {
+      const params = navigation.state.params;
+
+      if (params && paramName in params) {
+        return params[paramName];
+      }
+
+      return defaultValue;
+    },
+
+    push: (routeName, params, action) =>
+      navigation.dispatch(
+        NavigationActions.push({ routeName, params, action })
+      ),
+
+    replace: (routeName, params, action) =>
+      navigation.dispatch(
+        NavigationActions.replace({
+          routeName,
+          params,
+          action,
+          key: navigation.state.key,
+        })
+      ),
   };
 }
