@@ -1,25 +1,9 @@
-/* @flow */
-
 import React from 'react';
-import { BackHandler, Linking } from './PlatformHelpers';
+import { Linking } from 'react-native';
+import { BackHandler } from './PlatformHelpers';
 import NavigationActions from './NavigationActions';
 import addNavigationHelpers from './addNavigationHelpers';
 import invariant from './utils/invariant';
-
-import type {
-  NavigationAction,
-  NavigationState,
-  NavigationScreenProp,
-  NavigationNavigator,
-  PossiblyDeprecatedNavigationAction,
-  NavigationInitAction,
-  NavigationContainerProps,
-  NavigationContainer,
-} from './TypeDefinition';
-
-type State<NavState> = {
-  nav: ?NavState,
-};
 
 /**
  * Create an HOC that injects the navigation and manages the navigation state
@@ -27,38 +11,47 @@ type State<NavState> = {
  * This allows to use e.g. the StackNavigator and TabNavigator as root-level
  * components.
  */
-export default function createNavigationContainer<S: NavigationState, O: {}>(
-  // Let the NavigationNavigator props flowwwww
-  Component: NavigationNavigator<S, O, *>
-): NavigationContainer<S, O, *> {
-  class NavigationContainer extends React.Component<
-    NavigationContainerProps<S, O>,
-    State<S>
-  > {
-    subs: ?{
-      remove: () => void,
-    } = null;
+export default function createNavigationContainer(Component) {
+  class NavigationContainer extends React.Component {
+    subs = null;
 
     static router = Component.router;
     static navigationOptions = null;
 
-    constructor(props: NavigationContainerProps<S, O>) {
+    _actionEventSubscribers = new Set();
+
+    constructor(props) {
       super(props);
 
       this._validateProps(props);
 
+      this._initialAction = NavigationActions.init();
+
+      if (this._isStateful()) {
+        this.subs = BackHandler.addEventListener('hardwareBackPress', () => {
+          if (!this._isMounted) {
+            this.subs && this.subs.remove();
+          } else {
+            // dispatch returns true if the action results in a state change,
+            // and false otherwise. This maps well to what BackHandler expects
+            // from a callback -- true if handled, false if not handled
+            return this.dispatch(NavigationActions.back());
+          }
+        });
+      }
+
       this.state = {
         nav: this._isStateful()
-          ? Component.router.getStateForAction(NavigationActions.init())
+          ? Component.router.getStateForAction(this._initialAction)
           : null,
       };
     }
 
-    _isStateful(): boolean {
+    _isStateful() {
       return !this.props.navigation;
     }
 
-    _validateProps(props: NavigationContainerProps<S, O>) {
+    _validateProps(props) {
       if (this._isStateful()) {
         return;
       }
@@ -79,12 +72,14 @@ export default function createNavigationContainer<S: NavigationState, O: {}>(
       }
     }
 
-    _urlToPathAndParams(url: string) {
+    _urlToPathAndParams(url) {
       const params = {};
       const delimiter = this.props.uriPrefix || '://';
       let path = url.split(delimiter)[1];
       if (typeof path === 'undefined') {
         path = url;
+      } else if (path === '') {
+        path = '/';
       }
       return {
         path,
@@ -92,7 +87,7 @@ export default function createNavigationContainer<S: NavigationState, O: {}>(
       };
     }
 
-    _handleOpenURL = ({ url }: { url: string }) => {
+    _handleOpenURL = ({ url }) => {
       const parsedUrl = this._urlToPathAndParams(url);
       if (parsedUrl) {
         const { path, params } = parsedUrl;
@@ -103,11 +98,7 @@ export default function createNavigationContainer<S: NavigationState, O: {}>(
       }
     };
 
-    _onNavigationStateChange(
-      prevNav: NavigationState,
-      nav: NavigationState,
-      action: NavigationAction
-    ) {
+    _onNavigationStateChange(prevNav, nav, action) {
       if (
         typeof this.props.onNavigationStateChange === 'undefined' &&
         this._isStateful() &&
@@ -136,7 +127,7 @@ export default function createNavigationContainer<S: NavigationState, O: {}>(
       }
     }
 
-    componentWillReceiveProps(nextProps: NavigationContainerProps<S, O>) {
+    componentWillReceiveProps(nextProps) {
       this._validateProps(nextProps);
     }
 
@@ -148,31 +139,34 @@ export default function createNavigationContainer<S: NavigationState, O: {}>(
     }
 
     componentDidMount() {
+      this._isMounted = true;
       if (!this._isStateful()) {
         return;
       }
 
-      this.subs = BackHandler.addEventListener('hardwareBackPress', () =>
-        this.dispatch(NavigationActions.back())
-      );
-
       Linking.addEventListener('url', this._handleOpenURL);
 
-      Linking.getInitialURL().then(
-        (url: ?string) => url && this._handleOpenURL({ url })
+      Linking.getInitialURL().then(url => url && this._handleOpenURL({ url }));
+
+      this._actionEventSubscribers.forEach(subscriber =>
+        subscriber({
+          type: 'action',
+          action: this._initialAction,
+          state: this.state.nav,
+          lastState: null,
+        })
       );
     }
 
     componentWillUnmount() {
+      this._isMounted = false;
       Linking.removeEventListener('url', this._handleOpenURL);
       this.subs && this.subs.remove();
     }
 
     // Per-tick temporary storage for state.nav
-    _nav: ?S;
 
-    dispatch = (inputAction: PossiblyDeprecatedNavigationAction) => {
-      const action = NavigationActions.mapDeprecatedActionAndWarn(inputAction);
+    dispatch = action => {
       if (!this._isStateful()) {
         return false;
       }
@@ -180,18 +174,29 @@ export default function createNavigationContainer<S: NavigationState, O: {}>(
       const oldNav = this._nav;
       invariant(oldNav, 'should be set in constructor if stateful');
       const nav = Component.router.getStateForAction(action, oldNav);
+      const dispatchActionEvents = () => {
+        this._actionEventSubscribers.forEach(subscriber =>
+          subscriber({
+            type: 'action',
+            action,
+            state: nav,
+            lastState: oldNav,
+          })
+        );
+      };
       if (nav && nav !== oldNav) {
         // Cache updates to state.nav during the tick to ensure that subsequent calls will not discard this change
         this._nav = nav;
-        this.setState({ nav }, () =>
-          this._onNavigationStateChange(oldNav, nav, action)
-        );
+        this.setState({ nav }, () => {
+          this._onNavigationStateChange(oldNav, nav, action);
+          dispatchActionEvents();
+        });
         return true;
+      } else {
+        dispatchActionEvents();
       }
       return false;
     };
-
-    _navigation: ?NavigationScreenProp<NavigationState>;
 
     render() {
       let navigation = this.props.navigation;
@@ -202,6 +207,17 @@ export default function createNavigationContainer<S: NavigationState, O: {}>(
           this._navigation = addNavigationHelpers({
             dispatch: this.dispatch,
             state: nav,
+            addListener: (eventName, handler) => {
+              if (eventName !== 'action') {
+                return { remove: () => {} };
+              }
+              this._actionEventSubscribers.add(handler);
+              return {
+                remove: () => {
+                  this._actionEventSubscribers.delete(handler);
+                },
+              };
+            },
           });
         }
         navigation = this._navigation;
