@@ -1,7 +1,6 @@
-/* @flow */
-
-import * as React from 'react';
+import React from 'react';
 import { Dimensions, Platform, ScrollView } from 'react-native';
+import SafeAreaView from 'react-native-safe-area-view';
 
 import createNavigator from './createNavigator';
 import createNavigationContainer from '../createNavigationContainer';
@@ -9,36 +8,12 @@ import TabRouter from '../routers/TabRouter';
 import DrawerScreen from '../views/Drawer/DrawerScreen';
 import DrawerView from '../views/Drawer/DrawerView';
 import DrawerItems from '../views/Drawer/DrawerNavigatorItems';
-import SafeAreaView from '../views/SafeAreaView';
-
-import NavigatorTypes from './NavigatorTypes';
-
-import type { DrawerViewConfig } from '../views/Drawer/DrawerView';
-import type {
-  NavigationState,
-  NavigationRouteConfigMap,
-  NavigationTabRouterConfig,
-  NavigationDrawerScreenOptions,
-  NavigationNavigatorProps,
-} from '../TypeDefinition';
-
-export type DrawerNavigatorConfig = {
-  containerConfig?: void,
-} & NavigationTabRouterConfig &
-  DrawerViewConfig;
 
 // A stack navigators props are the intersection between
 // the base navigator props (navgiation, screenProps, etc)
 // and the view's props
-type DrawerNavigatorProps = NavigationNavigatorProps<
-  NavigationDrawerScreenOptions,
-  NavigationState
-> &
-  React.ElementProps<typeof DrawerView>;
 
-const defaultContentComponent = (
-  props: React.ElementProps<typeof DrawerItems>
-) => (
+const defaultContentComponent = props => (
   <ScrollView alwaysBounceVertical={false}>
     <SafeAreaView forceInset={{ top: 'always', horizontal: 'never' }}>
       <DrawerItems {...props} />
@@ -63,15 +38,15 @@ const DefaultDrawerConfig = {
     return Math.min(smallerAxisSize - appBarHeight, maxWidth);
   },
   contentComponent: defaultContentComponent,
+  drawerOpenRoute: 'DrawerOpen',
+  drawerCloseRoute: 'DrawerClose',
+  drawerToggleRoute: 'DrawerToggle',
   drawerPosition: 'left',
   drawerBackgroundColor: 'white',
   useNativeAnimations: true,
 };
 
-const DrawerNavigator = (
-  routeConfigs: NavigationRouteConfigMap,
-  config: DrawerNavigatorConfig = {}
-) => {
+const DrawerNavigator = (routeConfigs, config = {}) => {
   const mergedConfig = { ...DefaultDrawerConfig, ...config };
   const {
     containerConfig,
@@ -82,52 +57,49 @@ const DrawerNavigator = (
     drawerPosition,
     useNativeAnimations,
     drawerBackgroundColor,
+    drawerOpenRoute,
+    drawerCloseRoute,
+    drawerToggleRoute,
     ...tabsConfig
   } = mergedConfig;
 
   const contentRouter = TabRouter(routeConfigs, tabsConfig);
-
   const drawerRouter = TabRouter(
     {
-      DrawerClose: {
-        screen: createNavigator(
-          contentRouter,
-          routeConfigs,
-          config,
-          NavigatorTypes.DRAWER
-        )((props: React.ElementProps<typeof DrawerScreen>) => (
+      [drawerCloseRoute]: {
+        screen: createNavigator(contentRouter, routeConfigs, config)(props => (
           <DrawerScreen {...props} />
         )),
       },
-      DrawerOpen: {
+      [drawerOpenRoute]: {
         screen: () => null,
       },
-      DrawerToggle: {
+      [drawerToggleRoute]: {
         screen: () => null,
       },
     },
     {
-      initialRouteName: 'DrawerClose',
+      initialRouteName: drawerCloseRoute,
     }
   );
 
-  const navigator = createNavigator(
-    drawerRouter,
-    routeConfigs,
-    config,
-    NavigatorTypes.DRAWER
-  )((props: DrawerNavigatorProps) => (
-    <DrawerView
-      {...props}
-      drawerBackgroundColor={drawerBackgroundColor}
-      drawerLockMode={drawerLockMode}
-      useNativeAnimations={useNativeAnimations}
-      drawerWidth={drawerWidth}
-      contentComponent={contentComponent}
-      contentOptions={contentOptions}
-      drawerPosition={drawerPosition}
-    />
-  ));
+  const navigator = createNavigator(drawerRouter, routeConfigs, config)(
+    props => (
+      <DrawerView
+        {...props}
+        drawerBackgroundColor={drawerBackgroundColor}
+        drawerLockMode={drawerLockMode}
+        useNativeAnimations={useNativeAnimations}
+        drawerWidth={drawerWidth}
+        contentComponent={contentComponent}
+        contentOptions={contentOptions}
+        drawerPosition={drawerPosition}
+        drawerOpenRoute={drawerOpenRoute}
+        drawerCloseRoute={drawerCloseRoute}
+        drawerToggleRoute={drawerToggleRoute}
+      />
+    )
+  );
 
   return createNavigationContainer(navigator);
 };
