@@ -1,38 +1,22 @@
-/* @flow */
-
-import * as React from 'react';
+import React from 'react';
 import createNavigationContainer from '../createNavigationContainer';
 import createNavigator from './createNavigator';
 import CardStackTransitioner from '../views/CardStack/CardStackTransitioner';
 import StackRouter from '../routers/StackRouter';
-import NavigatorTypes from './NavigatorTypes';
-
-import type {
-  NavigationRouteConfigMap,
-  StackNavigatorConfig,
-  NavigationState,
-  NavigationStackScreenOptions,
-  NavigationNavigatorProps,
-} from '../TypeDefinition';
+import NavigationActions from '../NavigationActions';
 
 // A stack navigators props are the intersection between
 // the base navigator props (navgiation, screenProps, etc)
 // and the view's props
-type StackNavigatorProps = NavigationNavigatorProps<
-  NavigationStackScreenOptions,
-  NavigationState
-> &
-  React.ElementProps<typeof CardStackTransitioner>;
 
-export default (
-  routeConfigMap: NavigationRouteConfigMap,
-  stackConfig: StackNavigatorConfig = {}
-) => {
+export default (routeConfigMap, stackConfig = {}) => {
   const {
+    initialRouteKey,
     initialRouteName,
     initialRouteParams,
     paths,
     headerMode,
+    headerTransitionPreset,
     mode,
     cardStyle,
     transitionConfig,
@@ -42,6 +26,7 @@ export default (
   } = stackConfig;
 
   const stackRouterConfig = {
+    initialRouteKey,
     initialRouteName,
     initialRouteParams,
     paths,
@@ -51,22 +36,24 @@ export default (
   const router = StackRouter(routeConfigMap, stackRouterConfig);
 
   // Create a navigator with CardStackTransitioner as the view
-  const navigator = createNavigator(
-    router,
-    routeConfigMap,
-    stackConfig,
-    NavigatorTypes.STACK
-  )((props: StackNavigatorProps) => (
-    <CardStackTransitioner
-      {...props}
-      headerMode={headerMode}
-      mode={mode}
-      cardStyle={cardStyle}
-      transitionConfig={transitionConfig}
-      onTransitionStart={onTransitionStart}
-      onTransitionEnd={onTransitionEnd}
-    />
-  ));
+  const navigator = createNavigator(router, routeConfigMap, stackConfig)(
+    props => (
+      <CardStackTransitioner
+        {...props}
+        headerMode={headerMode}
+        headerTransitionPreset={headerTransitionPreset}
+        mode={mode}
+        cardStyle={cardStyle}
+        transitionConfig={transitionConfig}
+        onTransitionStart={onTransitionStart}
+        onTransitionEnd={(lastTransition, transition) => {
+          const { state, dispatch } = props.navigation;
+          dispatch(NavigationActions.completeTransition({ key: state.key }));
+          onTransitionEnd && onTransitionEnd(lastTransition, transition);
+        }}
+      />
+    )
+  );
 
   return createNavigationContainer(navigator);
 };
