@@ -6,9 +6,9 @@ import DrawerNavigator from '../DrawerNavigator';
 
 class HomeScreen extends Component {
   static navigationOptions = ({ navigation }) => ({
-    title: `Welcome ${navigation.state.params
-      ? navigation.state.params.user
-      : 'anonymous'}`,
+    title: `Welcome ${
+      navigation.state.params ? navigation.state.params.user : 'anonymous'
+    }`,
     gesturesEnabled: true,
   });
 
