@@ -1,7 +1,3 @@
-/*
- * @noflow - get/set properties not yet supported by flow. also `...require(x)` is broken #6560135
- */
-
 /* eslint global-require: 0 */
 
 module.exports = {
@@ -35,5 +31,8 @@ module.exports = {
   // HOCs
   get withNavigation() {
     return require('./views/withNavigation').default;
+  },
+  get withNavigationFocus() {
+    return require('./views/withNavigationFocus').default;
   },
 };
