@@ -1,27 +1,11 @@
-/*
- * @flow
- */
-
 import invariant from '../utils/invariant';
 
 import getScreenForRouteName from './getScreenForRouteName';
 import addNavigationHelpers from '../addNavigationHelpers';
 import validateScreenOptions from './validateScreenOptions';
+import getChildEventSubscriber from '../getChildEventSubscriber';
 
-import type {
-  NavigationScreenProp,
-  NavigationRoute,
-  NavigationStateRoute,
-  NavigationRouteConfigMap,
-  NavigationScreenConfig,
-  NavigationScreenConfigProps,
-} from '../TypeDefinition';
-
-function applyConfig<T: {}>(
-  configurer: ?NavigationScreenConfig<T>,
-  navigationOptions: any,
-  configProps: NavigationScreenConfigProps
-): * {
+function applyConfig(configurer, navigationOptions, configProps) {
   if (typeof configurer === 'function') {
     return {
       ...navigationOptions,
@@ -40,10 +24,10 @@ function applyConfig<T: {}>(
   return navigationOptions;
 }
 
-export default (
-  routeConfigs: NavigationRouteConfigMap,
-  navigatorScreenConfig?: NavigationScreenConfig<*>
-) => (navigation: NavigationScreenProp<NavigationRoute>, screenProps: *) => {
+export default (routeConfigs, navigatorScreenConfig) => (
+  navigation,
+  screenProps
+) => {
   const { state, dispatch } = navigation;
   const route = state;
 
@@ -58,8 +42,7 @@ export default (
 
   const router = Component.router;
   if (router) {
-    // $FlowFixMe
-    const { routes, index } = (route: NavigationStateRoute);
+    const { routes, index } = route;
     if (!route || !routes || index == null) {
       throw new Error(
         `Expect nav state to have routes and index, ${JSON.stringify(route)}`
@@ -69,6 +52,10 @@ export default (
     const childNavigation = addNavigationHelpers({
       state: childRoute,
       dispatch,
+      addListener: getChildEventSubscriber(
+        navigation.addListener,
+        childRoute.key
+      ),
     });
     outputConfig = router.getScreenOptions(childNavigation, screenProps);
   }
