@@ -1,26 +1,12 @@
-/* @flow */
-
-import type {
-  NavigationSceneRendererProps,
-  NavigationScene,
-  SceneIndicesForInterpolationInputRange,
-} from '../TypeDefinition';
-
-function getSceneIndicesForInterpolationInputRange(
-  props: NavigationSceneRendererProps
-): SceneIndicesForInterpolationInputRange | null {
+function getSceneIndicesForInterpolationInputRange(props) {
   const { scene, scenes } = props;
   const index = scene.index;
   const lastSceneIndexInScenes = scenes.length - 1;
   const isBack = !scenes[lastSceneIndexInScenes].isActive;
 
   if (isBack) {
-    const currentSceneIndexInScenes = scenes.findIndex(
-      (item: NavigationScene) => item === scene
-    );
-    const targetSceneIndexInScenes = scenes.findIndex(
-      (item: NavigationScene) => item.isActive
-    );
+    const currentSceneIndexInScenes = scenes.findIndex(item => item === scene);
+    const targetSceneIndexInScenes = scenes.findIndex(item => item.isActive);
     const targetSceneIndex = scenes[targetSceneIndexInScenes].index;
     const lastSceneIndex = scenes[lastSceneIndexInScenes].index;
 
