@@ -7,7 +7,7 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  *
  * @typechecks
- * @flow
+ *
  */
 
 /*eslint-disable no-self-compare */
@@ -20,7 +20,7 @@ const hasOwnProperty = Object.prototype.hasOwnProperty;
  * inlined Object.is polyfill to avoid requiring consumers ship their own
  * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/is
  */
-function is(x: mixed, y: mixed): boolean {
+function is(x, y) {
   // SameValue algorithm
   if (x === y) {
     // Steps 1-5, 7-10
@@ -38,7 +38,7 @@ function is(x: mixed, y: mixed): boolean {
  * when any key has values which are not strictly equal between the arguments.
  * Returns true when the values of all keys are strictly equal.
  */
-function shallowEqual(objA: mixed, objB: mixed): boolean {
+function shallowEqual(objA, objB) {
   if (is(objA, objB)) {
     return true;
   }
