@@ -1,12 +1,4 @@
-/* @flow */
-
 import { I18nManager } from 'react-native';
-
-import type {
-  NavigationSceneRendererProps,
-  AnimatedViewStyleProp,
-} from '../../TypeDefinition';
-
 import getSceneIndicesForInterpolationInputRange from '../../utils/getSceneIndicesForInterpolationInputRange';
 
 /**
@@ -27,9 +19,7 @@ import getSceneIndicesForInterpolationInputRange from '../../utils/getSceneIndic
 /**
  * Render the initial style when the initial layout isn't measured yet.
  */
-function forInitial(
-  props: NavigationSceneRendererProps
-): AnimatedViewStyleProp {
+function forInitial(props) {
   const { navigation, scene } = props;
 
   const focused = navigation.state.index === scene.index;
@@ -45,9 +35,7 @@ function forInitial(
 /**
  * Standard iOS-style slide in from the right.
  */
-function forHorizontal(
-  props: NavigationSceneRendererProps
-): AnimatedViewStyleProp {
+function forHorizontal(props) {
   const { layout, position, scene } = props;
 
   if (!layout.isMeasured) {
@@ -61,15 +49,15 @@ function forHorizontal(
   const index = scene.index;
   const opacity = position.interpolate({
     inputRange: [first, first + 0.01, index, last - 0.01, last],
-    outputRange: ([0, 1, 1, 0.85, 0]: Array<number>),
+    outputRange: [0, 1, 1, 0.85, 0],
   });
 
   const width = layout.initWidth;
   const translateX = position.interpolate({
-    inputRange: ([first, index, last]: Array<number>),
+    inputRange: [first, index, last],
     outputRange: I18nManager.isRTL
-      ? ([-width, 0, width * 0.3]: Array<number>)
-      : ([width, 0, width * -0.3]: Array<number>),
+      ? [-width, 0, width * 0.3]
+      : [width, 0, width * -0.3],
   });
   const translateY = 0;
 
@@ -82,9 +70,7 @@ function forHorizontal(
 /**
  * Standard iOS-style slide in from the bottom (used for modals).
  */
-function forVertical(
-  props: NavigationSceneRendererProps
-): AnimatedViewStyleProp {
+function forVertical(props) {
   const { layout, position, scene } = props;
 
   if (!layout.isMeasured) {
@@ -98,13 +84,13 @@ function forVertical(
   const index = scene.index;
   const opacity = position.interpolate({
     inputRange: [first, first + 0.01, index, last - 0.01, last],
-    outputRange: ([0, 1, 1, 0.85, 0]: Array<number>),
+    outputRange: [0, 1, 1, 0.85, 0],
   });
 
   const height = layout.initHeight;
   const translateY = position.interpolate({
-    inputRange: ([first, index, last]: Array<number>),
-    outputRange: ([height, 0, 0]: Array<number>),
+    inputRange: [first, index, last],
+    outputRange: [height, 0, 0],
   });
   const translateX = 0;
 
@@ -117,9 +103,7 @@ function forVertical(
 /**
  * Standard Android-style fade in from the bottom.
  */
-function forFadeFromBottomAndroid(
-  props: NavigationSceneRendererProps
-): AnimatedViewStyleProp {
+function forFadeFromBottomAndroid(props) {
   const { layout, position, scene } = props;
 
   if (!layout.isMeasured) {
@@ -131,16 +115,16 @@ function forFadeFromBottomAndroid(
 
   const { first, last } = interpolate;
   const index = scene.index;
-  const inputRange = ([first, index, last - 0.01, last]: Array<number>);
+  const inputRange = [first, index, last - 0.01, last];
 
   const opacity = position.interpolate({
     inputRange,
-    outputRange: ([0, 1, 1, 0]: Array<number>),
+    outputRange: [0, 1, 1, 0],
   });
 
   const translateY = position.interpolate({
     inputRange,
-    outputRange: ([50, 0, 0, 0]: Array<number>),
+    outputRange: [50, 0, 0, 0],
   });
   const translateX = 0;
 
@@ -153,7 +137,7 @@ function forFadeFromBottomAndroid(
 /**
  *  fadeIn and fadeOut
  */
-function forFade(props: NavigationSceneRendererProps): AnimatedViewStyleProp {
+function forFade(props) {
   const { layout, position, scene } = props;
 
   if (!layout.isMeasured) {
@@ -166,8 +150,8 @@ function forFade(props: NavigationSceneRendererProps): AnimatedViewStyleProp {
   const { first, last } = interpolate;
   const index = scene.index;
   const opacity = position.interpolate({
-    inputRange: ([first, index, last]: Array<number>),
-    outputRange: ([0, 1, 1]: Array<number>),
+    inputRange: [first, index, last],
+    outputRange: [0, 1, 1],
   });
 
   return {
@@ -175,7 +159,7 @@ function forFade(props: NavigationSceneRendererProps): AnimatedViewStyleProp {
   };
 }
 
-function canUseNativeDriver(): boolean {
+function canUseNativeDriver() {
   // The native driver can be enabled for this interpolator animating
   // opacity, translateX, and translateY is supported by the native animation
   // driver on iOS and Android.
