@@ -1,6 +1,4 @@
-/* @flow */
-
-import * as React from 'react';
+import React from 'react';
 import { NativeModules } from 'react-native';
 
 import CardStack from './CardStack';
@@ -8,38 +6,10 @@ import CardStackStyleInterpolator from './CardStackStyleInterpolator';
 import Transitioner from '../Transitioner';
 import TransitionConfigs from './TransitionConfigs';
 
-import type {
-  NavigationSceneRenderer,
-  NavigationScreenProp,
-  NavigationStackScreenOptions,
-  NavigationState,
-  NavigationTransitionProps,
-  NavigationNavigatorProps,
-  NavigationRouter,
-  HeaderMode,
-  ViewStyleProp,
-  TransitionConfig,
-} from '../../TypeDefinition';
-
 const NativeAnimatedModule =
   NativeModules && NativeModules.NativeAnimatedModule;
 
-type Props = {
-  headerMode: HeaderMode,
-  mode: 'card' | 'modal',
-  router: NavigationRouter<NavigationState, NavigationStackScreenOptions>,
-  cardStyle?: ViewStyleProp,
-  onTransitionStart?: () => void,
-  onTransitionEnd?: () => void,
-  /**
-   * Optional custom animation when transitioning between screens.
-   */
-  transitionConfig?: () => TransitionConfig,
-} & NavigationNavigatorProps<NavigationStackScreenOptions, NavigationState>;
-
-class CardStackTransitioner extends React.Component<Props> {
-  _render: NavigationSceneRenderer;
-
+class CardStackTransitioner extends React.Component {
   static defaultProps = {
     mode: 'card',
   };
@@ -58,13 +28,12 @@ class CardStackTransitioner extends React.Component<Props> {
 
   _configureTransition = (
     // props for the new screen
-    transitionProps: NavigationTransitionProps,
+    transitionProps,
     // props for the old screen
-    prevTransitionProps: ?NavigationTransitionProps
+    prevTransitionProps
   ) => {
     const isModal = this.props.mode === 'modal';
     // Copy the object so we can assign useNativeDriver below
-    // (avoid Flow error, transitionSpec is of type NavigationTransitionSpec).
     const transitionSpec = {
       ...TransitionConfigs.getTransitionConfig(
         this.props.transitionConfig,
@@ -84,10 +53,11 @@ class CardStackTransitioner extends React.Component<Props> {
     return transitionSpec;
   };
 
-  _render = (props: NavigationTransitionProps): React.Node => {
+  _render = (props, prevProps) => {
     const {
       screenProps,
       headerMode,
+      headerTransitionPreset,
       mode,
       router,
       cardStyle,
@@ -97,11 +67,13 @@ class CardStackTransitioner extends React.Component<Props> {
       <CardStack
         screenProps={screenProps}
         headerMode={headerMode}
+        headerTransitionPreset={headerTransitionPreset}
         mode={mode}
         router={router}
         cardStyle={cardStyle}
         transitionConfig={transitionConfig}
-        {...props}
+        transitionProps={props}
+        prevTransitionProps={prevProps}
       />
     );
   };
