@@ -1,38 +1,8 @@
-/* @flow */
-
-import * as React from 'react';
+import React from 'react';
 import { View, Text, Platform, StyleSheet } from 'react-native';
+import SafeAreaView from 'react-native-safe-area-view';
 
-import SafeAreaView from '../SafeAreaView';
 import TouchableItem from '../TouchableItem';
-
-import type {
-  NavigationScreenProp,
-  NavigationState,
-  NavigationRoute,
-  ViewStyleProp,
-  TextStyleProp,
-} from '../../TypeDefinition';
-import type { DrawerScene, DrawerItem } from './DrawerView.js';
-
-type Props = {
-  navigation: NavigationScreenProp<NavigationState>,
-  items: Array<NavigationRoute>,
-  activeItemKey?: ?string,
-  activeTintColor?: string,
-  activeBackgroundColor?: string,
-  inactiveTintColor?: string,
-  inactiveBackgroundColor?: string,
-  getLabel: (scene: DrawerScene) => ?(React.Node | string),
-  renderIcon: (scene: DrawerScene) => ?React.Node,
-  onItemPress: (info: DrawerItem) => void,
-  itemsContainerForceInset?: Object,
-  itemsContainerStyle?: ViewStyleProp,
-  itemStyle?: ViewStyleProp,
-  labelStyle?: TextStyleProp,
-  iconContainerStyle?: ViewStyleProp,
-  drawerPosition: 'left' | 'right',
-};
 
 /**
  * Component that renders the navigation list in the drawer.
@@ -51,11 +21,13 @@ const DrawerNavigatorItems = ({
   itemsContainerStyle,
   itemStyle,
   labelStyle,
+  activeLabelStyle,
+  inactiveLabelStyle,
   iconContainerStyle,
   drawerPosition,
-}: Props) => (
+}) => (
   <View style={[styles.container, itemsContainerStyle]}>
-    {items.map((route: NavigationRoute, index: number) => {
+    {items.map((route, index) => {
       const focused = activeItemKey === route.key;
       const color = focused ? activeTintColor : inactiveTintColor;
       const backgroundColor = focused
@@ -64,6 +36,7 @@ const DrawerNavigatorItems = ({
       const scene = { route, index, focused, tintColor: color };
       const icon = renderIcon(scene);
       const label = getLabel(scene);
+      const extraLabelStyle = focused ? activeLabelStyle : inactiveLabelStyle;
       return (
         <TouchableItem
           key={route.key}
@@ -93,7 +66,9 @@ const DrawerNavigatorItems = ({
                 </View>
               ) : null}
               {typeof label === 'string' ? (
-                <Text style={[styles.label, { color }, labelStyle]}>
+                <Text
+                  style={[styles.label, { color }, labelStyle, extraLabelStyle]}
+                >
                   {label}
                 </Text>
               ) : (
