@@ -1,6 +1,4 @@
-/* @flow */
-
-import * as React from 'react';
+import React from 'react';
 import {
   I18nManager,
   Image,
@@ -10,36 +8,22 @@ import {
   StyleSheet,
 } from 'react-native';
 
-import type { LayoutEvent, TextStyleProp } from '../../TypeDefinition';
-
 import TouchableItem from '../TouchableItem';
 
-type Props = {
-  onPress?: () => void,
-  pressColorAndroid?: string,
-  title?: ?string,
-  titleStyle?: ?TextStyleProp,
-  tintColor?: ?string,
-  truncatedTitle?: ?string,
-  width?: ?number,
-};
-
-type State = {
-  initialTextWidth?: number,
-};
-
-class HeaderBackButton extends React.PureComponent<Props, State> {
+class HeaderBackButton extends React.PureComponent {
   static defaultProps = {
     pressColorAndroid: 'rgba(0, 0, 0, .32)',
     tintColor: Platform.select({
       ios: '#037aff',
     }),
     truncatedTitle: 'Back',
+    // eslint-disable-next-line global-require
+    buttonImage: require('../assets/back-icon.png'),
   };
 
   state = {};
 
-  _onTextLayout = (e: LayoutEvent) => {
+  _onTextLayout = e => {
     if (this.state.initialTextWidth) {
       return;
     }
@@ -50,6 +34,7 @@ class HeaderBackButton extends React.PureComponent<Props, State> {
 
   render() {
     const {
+      buttonImage,
       onPress,
       pressColorAndroid,
       width,
@@ -65,9 +50,6 @@ class HeaderBackButton extends React.PureComponent<Props, State> {
         : false;
 
     const backButtonTitle = renderTruncated ? truncatedTitle : title;
-
-    // eslint-disable-next-line global-require
-    const asset = require('../assets/back-icon.png');
 
     return (
       <TouchableItem
@@ -88,7 +70,7 @@ class HeaderBackButton extends React.PureComponent<Props, State> {
               !!title && styles.iconWithTitle,
               !!tintColor && { tintColor },
             ]}
-            source={asset}
+            source={buttonImage}
           />
           {Platform.OS === 'ios' &&
             typeof backButtonTitle === 'string' && (
@@ -125,7 +107,7 @@ const styles = StyleSheet.create({
       ? {
           height: 21,
           width: 13,
-          marginLeft: 10,
+          marginLeft: 9,
           marginRight: 22,
           marginVertical: 12,
           resizeMode: 'contain',
@@ -141,7 +123,7 @@ const styles = StyleSheet.create({
   iconWithTitle:
     Platform.OS === 'ios'
       ? {
-          marginRight: 5,
+          marginRight: 6,
         }
       : {},
 });
