@@ -1,23 +1,9 @@
-/* @flow */
-
-import * as React from 'react';
-
+import React from 'react';
 import { Text, View, Platform, StyleSheet, Animated } from 'react-native';
-
-type AnimatedTextStyleProp = $PropertyType<
-  $PropertyType<Animated.Text, 'props'>,
-  'style'
->;
-
-type Props = {
-  children: React.Node,
-  selectionColor?: string | number,
-  style?: AnimatedTextStyleProp,
-};
 
 const AnimatedText = Animated.Text;
 
-const HeaderTitle = ({ style, ...rest }: Props) => (
+const HeaderTitle = ({ style, ...rest }) => (
   <AnimatedText
     numberOfLines={1}
     {...rest}
@@ -29,7 +15,7 @@ const HeaderTitle = ({ style, ...rest }: Props) => (
 const styles = StyleSheet.create({
   title: {
     fontSize: Platform.OS === 'ios' ? 17 : 20,
-    fontWeight: Platform.OS === 'ios' ? '600' : '500',
+    fontWeight: Platform.OS === 'ios' ? '700' : '500',
     color: 'rgba(0, 0, 0, .9)',
     textAlign: Platform.OS === 'ios' ? 'center' : 'left',
     marginHorizontal: 16,
