@@ -1,27 +1,7 @@
-/* @flow */
-
-import * as React from 'react';
+import React from 'react';
 import { Animated, View, StyleSheet } from 'react-native';
 
-import type {
-  NavigationState,
-  NavigationScreenProp,
-  ViewStyleProp,
-} from '../../TypeDefinition';
-
-import type { TabScene } from './TabView';
-
-type Props = {
-  activeTintColor: string,
-  inactiveTintColor: string,
-  scene: TabScene,
-  position: Animated.Value,
-  navigation: NavigationScreenProp<NavigationState>,
-  renderIcon: (scene: TabScene) => React.Node,
-  style?: ViewStyleProp,
-};
-
-export default class TabBarIcon extends React.PureComponent<Props> {
+export default class TabBarIcon extends React.PureComponent {
   render() {
     const {
       position,
@@ -34,15 +14,16 @@ export default class TabBarIcon extends React.PureComponent<Props> {
     const { route, index } = scene;
     const { routes } = navigation.state;
     // Prepend '-1', so there are always at least 2 items in inputRange
-    const inputRange = [-1, ...routes.map((x: *, i: number) => i)];
+    const inputRange = [-1, ...routes.map((x, i) => i)];
     const activeOpacity = position.interpolate({
       inputRange,
-      outputRange: inputRange.map((i: number) => (i === index ? 1 : 0)),
+      outputRange: inputRange.map(i => (i === index ? 1 : 0)),
     });
     const inactiveOpacity = position.interpolate({
       inputRange,
-      outputRange: inputRange.map((i: number) => (i === index ? 0 : 1)),
+      outputRange: inputRange.map(i => (i === index ? 0 : 1)),
     });
+
     // We render the icon twice at the same position on top of each other:
     // active and inactive one, so we can fade between them.
     return (
@@ -74,11 +55,11 @@ const styles = StyleSheet.create({
     // active and inactive one, so we can fade between them:
     // Cover the whole iconContainer:
     position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    alignSelf: 'center',
     alignItems: 'center',
     justifyContent: 'center',
+    height: '100%',
+    width: '100%',
+    minWidth: 30,
   },
 });
