@@ -1,46 +1,9 @@
-/* @flow */
-
-import * as React from 'react';
+import React from 'react';
 import { Animated, StyleSheet } from 'react-native';
 import { TabBar } from 'react-native-tab-view';
 import TabBarIcon from './TabBarIcon';
 
-import type {
-  NavigationAction,
-  NavigationRoute,
-  NavigationScreenProp,
-  NavigationState,
-  ViewStyleProp,
-  TextStyleProp,
-} from '../../TypeDefinition';
-
-import type { TabScene } from './TabView';
-
-type Props = {
-  activeTintColor: string,
-  inactiveTintColor: string,
-  showIcon: boolean,
-  showLabel: boolean,
-  upperCaseLabel: boolean,
-  allowFontScaling: boolean,
-  position: Animated.Value,
-  navigation: NavigationScreenProp<NavigationState>,
-  jumpToIndex: (index: number) => void,
-  getLabel: (scene: TabScene) => ?(React.Node | string),
-  getOnPress: (
-    previousScene: NavigationRoute,
-    scene: TabScene
-  ) => ({
-    previousScene: NavigationRoute,
-    scene: TabScene,
-    jumpToIndex: (index: number) => void,
-  }) => void,
-  renderIcon: (scene: TabScene) => React.Element<*>,
-  labelStyle?: TextStyleProp,
-  iconStyle?: ViewStyleProp,
-};
-
-export default class TabBarTop extends React.PureComponent<Props> {
+export default class TabBarTop extends React.PureComponent {
   static defaultProps = {
     activeTintColor: '#fff',
     inactiveTintColor: '#fff',
@@ -50,9 +13,10 @@ export default class TabBarTop extends React.PureComponent<Props> {
     allowFontScaling: true,
   };
 
-  _renderLabel = (scene: TabScene) => {
+  _renderLabel = scene => {
     const {
       position,
+      tabBarPosition,
       navigation,
       activeTintColor,
       inactiveTintColor,
@@ -67,14 +31,13 @@ export default class TabBarTop extends React.PureComponent<Props> {
     const { index } = scene;
     const { routes } = navigation.state;
     // Prepend '-1', so there are always at least 2 items in inputRange
-    const inputRange = [-1, ...routes.map((x: *, i: number) => i)];
+    const inputRange = [-1, ...routes.map((x, i) => i)];
     const outputRange = inputRange.map(
-      (inputIndex: number) =>
-        inputIndex === index ? activeTintColor : inactiveTintColor
+      inputIndex => (inputIndex === index ? activeTintColor : inactiveTintColor)
     );
     const color = position.interpolate({
       inputRange,
-      outputRange: (outputRange: Array<string>),
+      outputRange: outputRange,
     });
 
     const tintColor = scene.focused ? activeTintColor : inactiveTintColor;
@@ -96,7 +59,7 @@ export default class TabBarTop extends React.PureComponent<Props> {
     return label;
   };
 
-  _renderIcon = (scene: TabScene) => {
+  _renderIcon = scene => {
     const {
       position,
       navigation,
@@ -122,13 +85,21 @@ export default class TabBarTop extends React.PureComponent<Props> {
     );
   };
 
-  _handleOnPress = (scene: TabScene) => {
-    const { getOnPress, jumpToIndex, navigation }: Props = this.props;
+  _handleOnPress = scene => {
+    const { getOnPress, jumpToIndex, navigation } = this.props;
     const previousScene = navigation.state.routes[navigation.state.index];
     const onPress = getOnPress(previousScene, scene);
 
     if (onPress) {
-      onPress({ previousScene, scene, jumpToIndex });
+      // To maintain the same API as `TabbarBottom`, we pass in a `defaultHandler`
+      // even though I don't believe in this case it should be any different
+      // than `jumpToIndex`.
+      onPress({
+        previousScene,
+        scene,
+        jumpToIndex,
+        defaultHandler: jumpToIndex,
+      });
     } else {
       jumpToIndex(scene.index);
     }
@@ -136,7 +107,7 @@ export default class TabBarTop extends React.PureComponent<Props> {
 
   render() {
     // TODO: Define full proptypes
-    const props: any = this.props;
+    const props = this.props;
 
     return (
       <TabBar
