@@ -1,94 +1,52 @@
-/* @flow */
+import React from 'react';
+import { View, StyleSheet, Platform } from 'react-native';
+import { TabViewAnimated, TabViewPagerPan } from 'react-native-tab-view';
+import SafeAreaView from 'react-native-safe-area-view';
 
-import * as React from 'react';
-import {View, StyleSheet, Platform} from 'react-native';
-import {TabViewAnimated, TabViewPagerPan} from 'react-native-tab-view';
-import type {Layout} from 'react-native-tab-view/src/TabViewTypeDefinitions';
-import SceneView from '../SceneView';
+import ResourceSavingSceneView from '../ResourceSavingSceneView';
 import withCachedChildNavigation from '../../withCachedChildNavigation';
-import SafeAreaView from '../SafeAreaView';
 
-import type {
-  NavigationScreenProp,
-  NavigationRoute,
-  NavigationState,
-  NavigationRouter,
-  NavigationTabScreenOptions,
-} from '../../TypeDefinition';
-
-export type TabViewConfig = {
-  tabBarComponent?: React.ComponentType<*>,
-  tabBarPosition?: 'top' | 'bottom',
-  tabBarOptions?: {},
-  swipeEnabled?: boolean,
-  animationEnabled?: boolean,
-  configureTransition?: (
-    currentTransitionProps: Object,
-    nextTransitionProps: Object
-  ) => Object,
-  lazy?: boolean,
-  initialLayout?: Layout,
-};
-
-export type TabScene = {
-  route: NavigationRoute,
-  focused: boolean,
-  index: number,
-  tintColor?: ?string,
-};
-
-type Props = {
-  tabBarComponent?: React.ComponentType<*>,
-  tabBarPosition?: 'top' | 'bottom',
-  tabBarOptions?: {},
-  swipeEnabled?: boolean,
-  animationEnabled?: boolean,
-  configureTransition?: (
-    currentTransitionProps: Object,
-    nextTransitionProps: Object
-  ) => Object,
-  lazy?: boolean,
-  initialLayout: Layout,
-
-  screenProps?: {},
-  navigation: NavigationScreenProp<NavigationState>,
-  router: NavigationRouter<NavigationState, NavigationTabScreenOptions>,
-  childNavigationProps: {
-    [key: string]: NavigationScreenProp<NavigationRoute>,
-  },
-};
-
-class TabView extends React.PureComponent<Props> {
+class TabView extends React.PureComponent {
   static defaultProps = {
+    lazy: true,
+    removedClippedSubviews: true,
     // fix for https://github.com/react-native-community/react-native-tab-view/issues/312
     initialLayout: Platform.select({
       android: {width: 1, height: 0},
     }),
   };
 
-  _handlePageChanged = (index: number) => {
-    const {navigation} = this.props;
+  _handlePageChanged = index => {
+    const { navigation } = this.props;
     navigation.navigate(navigation.state.routes[index].routeName);
   };
 
-  _renderScene = ({route}: any) => {
-    const {screenProps} = this.props;
+  _renderScene = ({ route }) => {
+    const { screenProps, navigation } = this.props;
+    const focusedIndex = navigation.state.index;
+    const focusedKey = navigation.state.routes[focusedIndex].key;
+    const key = route.key;
     const childNavigation = this.props.childNavigationProps[route.key];
     const TabComponent = this.props.router.getComponentForRouteName(
       route.routeName
     );
+
     return (
-      <View style={styles.page}>
-        <SceneView
-          screenProps={screenProps}
-          component={TabComponent}
-          navigation={childNavigation}
-        />
-      </View>
+      <ResourceSavingSceneView
+        lazy={this.props.lazy}
+        isFocused={focusedKey === key}
+        removeClippedSubViews={this.props.removeClippedSubviews}
+        animationEnabled={this.props.animationEnabled}
+        swipeEnabled={this.props.swipeEnabled}
+        screenProps={screenProps}
+        component={TabComponent}
+        navigation={this.props.navigation}
+        childNavigation={childNavigation}
+      />
     );
   };
 
-  _getLabel = ({route, tintColor, focused}: TabScene) => {
+  _getLabel = ({ route, tintColor, focused }) => {
     const options = this.props.router.getScreenOptions(
       this.props.childNavigationProps[route.key],
       this.props.screenProps || {}
@@ -107,7 +65,7 @@ class TabView extends React.PureComponent<Props> {
     return route.routeName;
   };
 
-  _getOnPress = (previousScene: TabScene, {route}: TabScene) => {
+  _getOnPress = (previousScene, { route }) => {
     const options = this.props.router.getScreenOptions(
       this.props.childNavigationProps[route.key],
       this.props.screenProps || {}
@@ -116,16 +74,18 @@ class TabView extends React.PureComponent<Props> {
     return options.tabBarOnPress;
   };
 
-  _getTestIDProps = ({route}: TabScene) => {
+  _getTestIDProps = ({ route, focused }) => {
     const options = this.props.router.getScreenOptions(
       this.props.childNavigationProps[route.key],
       this.props.screenProps || {}
     );
 
-    return options.tabBarTestIDProps;
+    return typeof options.tabBarTestIDProps === 'function'
+      ? options.tabBarTestIDProps({ focused })
+      : options.tabBarTestIDProps;
   };
 
-  _renderIcon = ({focused, route, tintColor}: TabScene) => {
+  _renderIcon = ({ focused, route, tintColor }) => {
     const options = this.props.router.getScreenOptions(
       this.props.childNavigationProps[route.key],
       this.props.screenProps || {}
@@ -138,7 +98,7 @@ class TabView extends React.PureComponent<Props> {
     return null;
   };
 
-  _renderTabBar = (props: *) => {
+  _renderTabBar = props => {
     const {
       tabBarOptions,
       tabBarComponent: TabBarComponent,
@@ -147,10 +107,12 @@ class TabView extends React.PureComponent<Props> {
     if (typeof TabBarComponent === 'undefined') {
       return null;
     }
+
     return (
       <TabBarComponent
         {...props}
         {...tabBarOptions}
+        tabBarPosition={this.props.tabBarPosition}
         screenProps={this.props.screenProps}
         navigation={this.props.navigation}
         getLabel={this._getLabel}
@@ -162,7 +124,7 @@ class TabView extends React.PureComponent<Props> {
     );
   };
 
-  _renderPager = (props: *) => <TabViewPagerPan {...props} />;
+  _renderPager = props => <TabViewPagerPan {...props} />;
 
   render() {
     const {
@@ -171,8 +133,6 @@ class TabView extends React.PureComponent<Props> {
       tabBarPosition,
       animationEnabled,
       configureTransition,
-      swipeEnabled,
-      lazy,
       initialLayout,
       screenProps,
     } = this.props;
@@ -190,6 +150,15 @@ class TabView extends React.PureComponent<Props> {
     const tabBarVisible =
       options.tabBarVisible == null ? true : options.tabBarVisible;
 
+    let swipeEnabled =
+      options.swipeEnabled == null
+        ? this.props.swipeEnabled
+        : options.swipeEnabled;
+
+    if (typeof swipeEnabled === 'function') {
+      swipeEnabled = swipeEnabled(state);
+    }
+
     if (tabBarComponent !== undefined && tabBarVisible) {
       if (tabBarPosition === 'bottom') {
         renderFooter = this._renderTabBar;
@@ -206,7 +175,6 @@ class TabView extends React.PureComponent<Props> {
     }
 
     const props = {
-      lazy,
       initialLayout,
       animationEnabled,
       configureTransition,
@@ -222,7 +190,6 @@ class TabView extends React.PureComponent<Props> {
       style: styles.container,
     };
 
-    // $FlowFixMe: mismatch with react-native-tab-view type
     return <TabViewAnimated {...props} />;
   }
 }
@@ -232,10 +199,5 @@ export default withCachedChildNavigation(TabView);
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-  },
-
-  page: {
-    flex: 1,
-    overflow: 'hidden',
   },
 });
