@@ -1,5 +1,3 @@
-/* @flow */
-
 /**
  * TouchableItem renders a touchable that looks native on both iOS and Android.
  *
@@ -9,28 +7,17 @@
  * On iOS you can pass the props of TouchableOpacity, on Android pass the props
  * of TouchableNativeFeedback.
  */
-import * as React from 'react';
+import React from 'react';
 import {
   Platform,
   TouchableNativeFeedback,
   TouchableOpacity,
   View,
 } from 'react-native';
-import type { ViewStyleProp } from '../TypeDefinition';
 
 const ANDROID_VERSION_LOLLIPOP = 21;
 
-type Props = {
-  onPress?: () => void,
-  delayPressIn?: number,
-  borderless?: boolean,
-  pressColor?: string,
-  activeOpacity?: number,
-  children?: React.Node,
-  style?: ViewStyleProp,
-};
-
-export default class TouchableItem extends React.Component<Props> {
+export default class TouchableItem extends React.Component {
   static defaultProps = {
     borderless: false,
     pressColor: 'rgba(0, 0, 0, .32)',
@@ -55,8 +42,8 @@ export default class TouchableItem extends React.Component<Props> {
           {...rest}
           style={null}
           background={TouchableNativeFeedback.Ripple(
-            this.props.pressColor || '',
-            this.props.borderless || false
+            this.props.pressColor,
+            this.props.borderless
           )}
         >
           <View style={style}>{React.Children.only(this.props.children)}</View>
