@@ -1,64 +1,18 @@
-/* @flow */
-
-import * as React from 'react';
-
+import React from 'react';
 import { Animated, Easing, StyleSheet, View } from 'react-native';
-
 import invariant from '../utils/invariant';
 
 import NavigationScenesReducer from './ScenesReducer';
 
-import type {
-  NavigationLayout,
-  NavigationScene,
-  NavigationState,
-  NavigationScreenProp,
-  NavigationTransitionProps,
-  NavigationTransitionSpec,
-} from '../TypeDefinition';
-
-type Props = {
-  configureTransition: (
-    transitionProps: NavigationTransitionProps,
-    prevTransitionProps: ?NavigationTransitionProps
-  ) => NavigationTransitionSpec,
-  navigation: NavigationScreenProp<NavigationState>,
-  onTransitionEnd?: (...args: Array<mixed>) => void,
-  onTransitionStart?: (...args: Array<mixed>) => void,
-  render: (
-    transitionProps: NavigationTransitionProps,
-    prevTransitionProps: ?NavigationTransitionProps
-  ) => React.Node,
-};
-
-type State = {
-  layout: NavigationLayout,
-  position: Animated.Value,
-  progress: Animated.Value,
-  scenes: Array<NavigationScene>,
-};
-
 // Used for all animations unless overriden
-const DefaultTransitionSpec = ({
+const DefaultTransitionSpec = {
   duration: 250,
   easing: Easing.inOut(Easing.ease),
   timing: Animated.timing,
-}: NavigationTransitionSpec);
+};
 
-class Transitioner extends React.Component<Props, State> {
-  _onLayout: (event: any) => void;
-  _onTransitionEnd: () => void;
-  _prevTransitionProps: ?NavigationTransitionProps;
-  _transitionProps: NavigationTransitionProps;
-  _isMounted: boolean;
-  _isTransitionRunning: boolean;
-  _queuedTransition: ?{
-    nextProps: Props,
-    nextScenes: Array<NavigationScene>,
-    indexHasChanged: boolean,
-  };
-
-  constructor(props: Props, context: any) {
+class Transitioner extends React.Component {
+  constructor(props, context) {
     super(props, context);
 
     // The initial layout isn't measured. Measured layout will be only available
@@ -85,20 +39,20 @@ class Transitioner extends React.Component<Props, State> {
     this._queuedTransition = null;
   }
 
-  componentWillMount(): void {
+  componentWillMount() {
     this._onLayout = this._onLayout.bind(this);
     this._onTransitionEnd = this._onTransitionEnd.bind(this);
   }
 
-  componentDidMount(): void {
+  componentDidMount() {
     this._isMounted = true;
   }
 
-  componentWillUnmount(): void {
+  componentWillUnmount() {
     this._isMounted = false;
   }
 
-  componentWillReceiveProps(nextProps: Props): void {
+  componentWillReceiveProps(nextProps) {
     const nextScenes = NavigationScenesReducer(
       this.state.scenes,
       nextProps.navigation.state,
@@ -119,11 +73,7 @@ class Transitioner extends React.Component<Props, State> {
     this._startTransition(nextProps, nextScenes, indexHasChanged);
   }
 
-  _startTransition(
-    nextProps: Props,
-    nextScenes: Array<NavigationScene>,
-    indexHasChanged: boolean
-  ) {
+  _startTransition(nextProps, nextScenes, indexHasChanged) {
     const nextState = {
       ...this.state,
       scenes: nextScenes,
@@ -172,12 +122,17 @@ class Transitioner extends React.Component<Props, State> {
 
     // update scenes and play the transition
     this._isTransitionRunning = true;
-    this.setState(nextState, () => {
-      nextProps.onTransitionStart &&
-        nextProps.onTransitionStart(
+    this.setState(nextState, async () => {
+      if (nextProps.onTransitionStart) {
+        const result = nextProps.onTransitionStart(
           this._transitionProps,
           this._prevTransitionProps
         );
+
+        if (result instanceof Promise) {
+          await result;
+        }
+      }
       Animated.parallel(animations).start(this._onTransitionEnd);
     });
   }
@@ -190,7 +145,7 @@ class Transitioner extends React.Component<Props, State> {
     );
   }
 
-  _onLayout(event: any): void {
+  _onLayout(event) {
     const { height, width } = event.nativeEvent.layout;
     if (
       this.state.layout.initWidth === width &&
@@ -217,23 +172,42 @@ class Transitioner extends React.Component<Props, State> {
     this.setState(nextState);
   }
 
-  _onTransitionEnd(): void {
+  _onTransitionEnd() {
     if (!this._isMounted) {
       return;
     }
     const prevTransitionProps = this._prevTransitionProps;
     this._prevTransitionProps = null;
 
+    const scenes = this.state.scenes.filter(isSceneNotStale);
+
     const nextState = {
       ...this.state,
-      scenes: this.state.scenes.filter(isSceneNotStale),
+      /**
+       * Array.prototype.filter creates a new instance of an array
+       * even if there were no elements removed. There are cases when
+       * `this.state.scenes` will have no stale scenes (typically when
+       * pushing a new route). As a result, components that rely on this prop
+       * might enter an unnecessary render cycle.
+       */
+      scenes:
+        this.state.scenes.length === scenes.length ? this.state.scenes : scenes,
     };
 
     this._transitionProps = buildTransitionProps(this.props, nextState);
 
-    this.setState(nextState, () => {
-      this.props.onTransitionEnd &&
-        this.props.onTransitionEnd(this._transitionProps, prevTransitionProps);
+    this.setState(nextState, async () => {
+      if (this.props.onTransitionEnd) {
+        const result = this.props.onTransitionEnd(
+          this._transitionProps,
+          prevTransitionProps
+        );
+
+        if (result instanceof Promise) {
+          await result;
+        }
+      }
+
       if (this._queuedTransition) {
         this._startTransition(
           this._queuedTransition.nextProps,
@@ -248,10 +222,7 @@ class Transitioner extends React.Component<Props, State> {
   }
 }
 
-function buildTransitionProps(
-  props: Props,
-  state: State
-): NavigationTransitionProps {
+function buildTransitionProps(props, state) {
   const { navigation } = props;
 
   const { layout, position, progress, scenes } = state;
@@ -271,11 +242,11 @@ function buildTransitionProps(
   };
 }
 
-function isSceneNotStale(scene: NavigationScene): boolean {
+function isSceneNotStale(scene) {
   return !scene.isStale;
 }
 
-function isSceneActive(scene: NavigationScene): boolean {
+function isSceneActive(scene) {
   return scene.isActive;
 }
 
