@@ -1,33 +1,27 @@
-/* @flow */
-
-import * as React from 'react';
+import React from 'react';
 import propTypes from 'prop-types';
 import hoistStatics from 'hoist-non-react-statics';
 
-import type { NavigationScreenProp, NavigationState } from '../TypeDefinition';
+export default function withNavigation(Component) {
+  class ComponentWithNavigation extends React.Component {
+    static displayName = `withNavigation(${Component.displayName ||
+      Component.name})`;
 
-type Context = {
-  navigation: NavigationScreenProp<NavigationState>,
-};
+    static contextTypes = {
+      navigation: propTypes.object.isRequired,
+    };
 
-type InjectedProps = {
-  navigation: NavigationScreenProp<NavigationState>,
-};
+    render() {
+      const { navigation } = this.context;
+      return (
+        <Component
+          {...this.props}
+          navigation={navigation}
+          ref={this.props.onRef}
+        />
+      );
+    }
+  }
 
-export default function withNavigation<T: {}>(
-  Component: React.ComponentType<T & InjectedProps>
-) {
-  const componentWithNavigation = (props: T, { navigation }: Context) => (
-    <Component {...props} navigation={navigation} />
-  );
-
-  // $FlowFixMe StatelessFunctionalComponent missing displayName Flow < 0.54.0
-  const displayName: string = Component.displayName || Component.name;
-  componentWithNavigation.displayName = `withNavigation(${displayName})`;
-
-  componentWithNavigation.contextTypes = {
-    navigation: propTypes.object.isRequired,
-  };
-
-  return hoistStatics(componentWithNavigation, Component);
+  return hoistStatics(ComponentWithNavigation, Component);
 }
