@@ -1,31 +1,11 @@
-// @flow
-
-import * as React from 'react';
+import React from 'react';
 import { Dimensions } from 'react-native';
 import hoistNonReactStatic from 'hoist-non-react-statics';
 
-type WindowDimensions = {
-  width: number,
-  height: number,
-};
+export const isOrientationLandscape = ({ width, height }) => width > height;
 
-type InjectedProps = {
-  isLandscape: boolean,
-};
-
-type State = {
-  isLandscape: boolean,
-};
-
-export const isOrientationLandscape = ({
-  width,
-  height,
-}: WindowDimensions): boolean => width > height;
-
-export default function<T: {}>(
-  WrappedComponent: React.ComponentType<T & InjectedProps>
-) {
-  class withOrientation extends React.Component<T, State> {
+export default function(WrappedComponent) {
+  class withOrientation extends React.Component {
     constructor() {
       super();
 
@@ -41,7 +21,7 @@ export default function<T: {}>(
       Dimensions.removeEventListener('change', this.handleOrientationChange);
     }
 
-    handleOrientationChange = ({ window }: { window: WindowDimensions }) => {
+    handleOrientationChange = ({ window }) => {
       const isLandscape = isOrientationLandscape(window);
       this.setState({ isLandscape });
     };
